requirements.txt
anyio==3.6.2
asgiref==3.6.0
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
fastapi==0.95.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.0.1
opentelemetry-api==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
prometheus-client==0.16.0
protobuf==4.23.0
pydantic==1.10.7
requests==2.30.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
wrapt==1.15.0
zipp==3.15.0
